msgpack-bind/src/test/java/com/lisb/msgpack/bind/marshaller/IntegerMarshallerTest.java
package com.lisb.msgpack.bind.marshaller;

import com.lisb.msgpack.bind.Marshaller;
import com.lisb.msgpack.bind.Unmarshaller;
import com.lisb.msgpack.bind.unmarshaller.IntegerUnmarshaller;
import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;
import org.msgpack.core.MessageBufferPacker;
import org.msgpack.core.MessagePack;

public class IntegerMarshallerTest {

    private static Marshaller<Integer> marshaller;
    private static Unmarshaller<Integer> unmarshaller;

    @BeforeClass
    public static void beforeClass() {
        marshaller = IntegerMarshaller.getInstance();
        unmarshaller = IntegerUnmarshaller.getInstance();
    }

    @Test
    public void testMarshalAndUnmarshal() throws Exception {
        final MessageBufferPacker packer = MessagePack.newDefaultBufferPacker();

        final Integer source = 6567;

        marshaller.marshal(packer, source);

        final Integer dest = unmarshaller.unmarshal(MessagePack.newDefaultUnpacker(packer.toByteArray()));

        Assert.assertEquals(source, dest);
    }

    @Test
    public void testNullBMarshalAndUnmarshal() throws Exception {
        final MessageBufferPacker packer = MessagePack.newDefaultBufferPacker();
        marshaller.marshal(packer, null);
        Assert.assertNull(unmarshaller.unmarshal(MessagePack.newDefaultUnpacker(packer.toByteArray())));
    }

}
